# backend/openapi.yaml
openapi: 3.0.3
info:
  title: SumbiTheses API
  version: 1.0.0
  description: API for managing thesis projects, students, supervisors, and schools
  contact:
    name: API Support
    email: support@sumbitheses.com

servers:
  - url: http://localhost:5000/api
    description: Development server

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Authorization header using the Bearer scheme

  schemas:
    Status:
      type: string
      enum: [draft, submitted, locked, public]
      description: Project status

    UserRole:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        description:
          type: string
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, name, created_at, updated_at]

    CreateUserRole:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
      required: [name]

    UpdateUserRole:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string

    School:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        domain:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, name, domain, created_at, updated_at]

    CreateSchool:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        domain:
          type: string
          minLength: 1
      required: [name, domain]

    UpdateSchool:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        domain:
          type: string
          minLength: 1

    SchoolUser:
      type: object
      properties:
        id:
          type: integer
          format: int64
        email:
          type: string
          format: email
        first_name:
          type: string
          nullable: true
        last_name:
          type: string
          nullable: true
        school_id:
          type: integer
          format: int64
        role_id:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, email, school_id, role_id, created_at, updated_at]

    CreateSchoolUser:
      type: object
      properties:
        email:
          type: string
          format: email
        first_name:
          type: string
        last_name:
          type: string
        school_id:
          type: integer
          format: int64
        role_id:
          type: integer
          format: int64
      required: [email, school_id, role_id]

    UpdateSchoolUser:
      type: object
      properties:
        email:
          type: string
          format: email
        first_name:
          type: string
        last_name:
          type: string
        school_id:
          type: integer
          format: int64
        role_id:
          type: integer
          format: int64

    ProjectStudent:
      type: object
      properties:
        id:
          type: integer
          format: int64
        project_id:
          type: integer
          format: int64
        student_id:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
        students:
          $ref: '#/components/schemas/SchoolUser'
      required: [id, project_id, student_id, created_at]

    Project:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        supervisor_id:
          type: integer
          format: int64
        opponent_id:
          type: integer
          format: int64
        subject:
          type: string
        description:
          type: string
          nullable: true
        main_document:
          type: string
          nullable: true
        locked_until:
          type: string
          format: date-time
          nullable: true
        status:
          $ref: '#/components/schemas/Status'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        project_students:
          type: array
          items:
            $ref: '#/components/schemas/ProjectStudent'
        school_users_projects_supervisor_idToschool_users:
          $ref: '#/components/schemas/SchoolUser'
        school_users_projects_opponent_idToschool_users:
          $ref: '#/components/schemas/SchoolUser'
      required: [id, title, supervisor_id, opponent_id, subject, status, created_at, updated_at]

    CreateProject:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        supervisor_id:
          type: integer
          format: int64
        opponent_id:
          type: integer
          format: int64
        subject:
          type: string
          minLength: 1
        description:
          type: string
        main_document:
          type: string
        locked_until:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/Status'
      required: [title, supervisor_id, opponent_id, subject]

    UpdateProject:
      type: object
      properties:
        title:
          type: string
          minLength: 1
        supervisor_id:
          type: integer
          format: int64
        opponent_id:
          type: integer
          format: int64
        subject:
          type: string
          minLength: 1
        description:
          type: string
        main_document:
          type: string
        locked_until:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/Status'

    AddStudentToProject:
      type: object
      properties:
        studentId:
          type: integer
          format: int64
      required: [studentId]

    UpdateProjectStudents:
      type: object
      properties:
        studentIds:
          type: array
          items:
            type: integer
            format: int64
      required: [studentIds]

    Review:
      type: object
      properties:
        id:
          type: integer
          format: int64
        project_id:
          type: integer
          format: int64
        reviewer_id:
          type: integer
          format: int64
        comments:
          type: string
        submitted_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, project_id, reviewer_id, comments, submitted_at, updated_at]

    CreateReview:
      type: object
      properties:
        project_id:
          type: integer
          format: int64
        reviewer_id:
          type: integer
          format: int64
        comments:
          type: string
          minLength: 1
      required: [project_id, reviewer_id, comments]

    UpdateReview:
      type: object
      properties:
        comments:
          type: string
          minLength: 1
      required: [comments]

    Attachment:
      type: object
      properties:
        id:
          type: integer
          format: int64
        project_id:
          type: integer
          format: int64
          nullable: true
        filename:
          type: string
        storage_path:
          type: string
        description:
          type: string
          nullable: true
        uploaded_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, filename, storage_path, uploaded_at, updated_at]

    CreateAttachment:
      type: object
      properties:
        project_id:
          type: integer
          format: int64
        filename:
          type: string
          minLength: 1
        storage_path:
          type: string
          minLength: 1
        description:
          type: string
      required: [filename, storage_path]

    UpdateAttachment:
      type: object
      properties:
        filename:
          type: string
          minLength: 1
        description:
          type: string

    ExternalLink:
      type: object
      properties:
        id:
          type: integer
          format: int64
        project_id:
          type: integer
          format: int64
          nullable: true
        url:
          type: string
          format: uri
          nullable: true
        title:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        added_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, added_at, updated_at]

    CreateExternalLink:
      type: object
      properties:
        project_id:
          type: integer
          format: int64
        url:
          type: string
          format: uri
        title:
          type: string
        description:
          type: string

    UpdateExternalLink:
      type: object
      properties:
        project_id:
          type: integer
          format: int64
        url:
          type: string
          format: uri
        title:
          type: string
        description:
          type: string

    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        role:
          type: string
        school_id:
          type: integer
          format: int64
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
      required: [id, email, role, created_at, updated_at]

    CreateUser:
      type: object
      properties:
        email:
          type: string
          format: email
        role:
          type: string
          default: student
        school_id:
          type: integer
          format: int64
      required: [email]

    UpdateUser:
      type: object
      properties:
        role:
          type: string
        school_id:
          type: integer
          format: int64

    # Error response schemas
    Error:
      type: object
      properties:
        error:
          type: string
      required: [error]

    ValidationError:
      type: object
      properties:
        error:
          type: string
        details:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
      required: [error]

  responses:
    UnauthorizedError:
      description: Authentication information is missing or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Unauthorized

    ForbiddenError:
      description: Access forbidden
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Forbidden

    NotFoundError:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Not found

    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ValidationError'
          example:
            error: Validation failed
            details:
              - field: title
                message: Title is required

    ServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: Internal server error

security:
  - bearerAuth: []

paths:
  # Projects endpoints
  /projects:
    get:
      summary: Get all projects
      tags: [Projects]
      responses:
        '200':
          description: List of projects
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/ServerError'

    post:
      summary: Create a new project
      tags: [Projects]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProject'
      responses:
        '201':
          description: Project created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/ServerError'

  /projects/{id}:
    get:
      summary: Get project by ID
      tags: [Projects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/ServerError'

    put:
      summary: Update project
      tags: [Projects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProject'
      responses:
        '200':
          description: Project updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/ServerError'

    delete:
      summary: Delete project
      tags: [Projects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Project deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/ServerError'

  /projects/{id}/students:
    get:
      summary: Get students assigned to project
      tags: [Projects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: List of students assigned to project
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SchoolUser'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/ServerError'

    post:
      summary: Add student to project
      tags: [Projects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddStudentToProject'
      responses:
        '201':
          description: Student added to project successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectStudent'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/ServerError'

    put:
      summary: Update all students assigned to project
      tags: [Projects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProjectStudents'
      responses:
        '200':
          description: Project students updated successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SchoolUser'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/ServerError'

  /projects/{id}/students/{studentId}:
    delete:
      summary: Remove student from project
      tags: [Projects]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: studentId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Student removed from project successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/ServerError'

  # Schools endpoints
  /schools:
    get:
      summary: Get all schools
      tags: [Schools]
      responses:
        '200':
          description: List of schools
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/School'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '500':
          $ref: '#/components/responses/ServerError'

    post:
      summary: Create a new school
      tags: [Schools]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSchool'
      responses:
        '201':
          description: School created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/School'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/ServerError'

  /schools/{id}:
    get:
      summary: Get school by ID
      tags: [Schools]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: School details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/School'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/ServerError'

    put:
      summary: Update school
      tags: [Schools]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSchool'
      responses:
        '200':
          description: School updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/School'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/ServerError'

    delete:
      summary: Delete school
      tags: [Schools]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: School deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/ServerError'

  /schools/{id}/students:
    get:
      summary: Get students by school ID
      tags: [Schools]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: List of students in school
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SchoolUser'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/ServerError'

  /schools/{id}/teachers:
    get:
      summary: Get teachers by school ID
      tags: [Schools]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: List of teachers in school
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SchoolUser'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/ServerError'

  # Users endpoints
  /users:
    get:
      summary: Get all users
      tags: [Users]
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '500':
          $ref: '#/components/responses/ServerError'

  /users/{id}:
    get:
      summary: Get user by ID
      tags: [Users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: User details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/ServerError'

    put:
      summary: Update user
      tags: [Users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUser'
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/ServerError'

    delete:
      summary: Delete user
      tags: [Users]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: User deleted successfully
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '500':
          $ref: '#/components/responses/ServerError'